+++ lib/cached/md5_cached.cmake
@@ -1,4 +1,4 @@
 set(lexer_lxx_md5_cached "32d69703b364c176cd579c08cd355b7a")
-set(parser_yxx_md5_cached "cfddc5f4a884d2a30f0cd07fc2d7fa29")
+set(parser_yxx_md5_cached "2b4a667843b7ae9fac80f95b0fc49e5d")
 set(regex_lexer_lxx_md5_cached "5b4c138ececa989083ac2ed39f8959c2")
 set(regex_parser_yxx_md5_cached "e44695248dcfad64a56982f126f8e4f5")